+++ t.go
@@ -5,7 +5,7 @@ import (
 	"runtime"
 )
 
-type T struct {
+type t struct {
 	name        string
 	tester      Tester
 	failed      bool
@@ -19,9 +19,9 @@ type subtest struct {
 	tester Tester
 }
 
-var _ TestingT = (*T)(nil)
+var _ TestingT = (*t)(nil)
 
-func (t *T) run() {
+func (t *t) run() {
 	defer func() {
 		// catch panics and mark test as failed
 		if err := recover(); err != nil {
@@ -35,47 +35,47 @@ func (t *T) run() {
 	t.tester(t)
 }
 
-func (t *T) Fail() {
+func (t *t) Fail() {
 	t.failed = true
 }
 
-func (t *T) FailNow() {
+func (t *t) FailNow() {
 	t.failed = true
 	runtime.Goexit()
 }
 
-func (t *T) Fatal(args ...interface{}) {
+func (t *t) Fatal(args ...interface{}) {
 	t.msgs = append(t.msgs, Msg{Msg: fmt.Sprintln(args...), Level: LevelError})
 	t.FailNow()
 }
 
-func (t *T) Fatalf(format string, args ...interface{}) {
+func (t *t) Fatalf(format string, args ...interface{}) {
 	t.msgs = append(t.msgs, Msg{Msg: fmt.Sprintf(format, args...), Level: LevelError})
 	t.FailNow()
 }
 
-func (t *T) Errorf(format string, args ...interface{}) {
+func (t *t) Errorf(format string, args ...interface{}) {
 	t.msgs = append(t.msgs, Msg{Msg: fmt.Sprintf(format, args...), Level: LevelError})
 	t.failed = true
 }
 
-func (t *T) Helper() {
+func (t *t) Helper() {
 	// nothing to do here, I think?
 }
 
-func (t *T) Log(args ...interface{}) {
+func (t *t) Log(args ...interface{}) {
 	t.msgs = append(t.msgs, Msg{Msg: fmt.Sprintln(args...), Level: LevelInfo})
 }
 
-func (t *T) Logf(format string, args ...interface{}) {
+func (t *t) Logf(format string, args ...interface{}) {
 	t.msgs = append(t.msgs, Msg{Msg: fmt.Sprintf(format, args...), Level: LevelInfo})
 }
 
-func (t *T) Name() string {
+func (t *t) Name() string {
 	return t.name
 }
 
-func (t *T) Run(name string, tester Tester) {
+func (t *t) Run(name string, tester Tester) {
 	t.subtests <- subtest{
 		name:   name,
 		tester: tester,
